+++ docs/firestore/example.mdx
@@ -1,0 +1,8 @@
+---
+title: Firestore Example
+hide_table_of_contents: true
+---
+
+<DartPad ref="dartpad" path="packages/cloud_firestore/cloud_firestore/example" />
+
+> View the [source code](https://github.com/FirebaseExtended/flutterfire/tree/master/packages/cloud_firestore/cloud_firestore/example).
